Move sample, head/tail, test_reductions.py
