+++ python/cudf/cudf/tests/test_groupby.py
@@ -1,9 +1,5 @@
 # Copyright (c) 2018-2025, NVIDIA CORPORATION.
 
-import collections
-import itertools
-import operator
-import string
 import textwrap
 from functools import partial
 
@@ -925,265 +921,6 @@ def test_groupby_group_keys(group_keys, by):
     actual = g_group[["B", "C"]].apply(lambda x: x / x.sum())
     expected = p_group[["B", "C"]].apply(lambda x: x / x.sum())
     assert_eq(actual, expected)
-
-
-@pytest.mark.parametrize("index_names", ["a", "b", "c", ["b", "c"]])
-def test_groupby_by_index_names(index_names):
-    gdf = cudf.DataFrame(
-        {"a": [1, 2, 3, 4], "b": ["a", "b", "a", "a"], "c": [1, 1, 2, 1]}
-    ).set_index(index_names)
-    pdf = gdf.to_pandas()
-
-    assert_groupby_results_equal(
-        pdf.groupby(index_names).min(), gdf.groupby(index_names).min()
-    )
-
-
-@pytest.mark.parametrize(
-    "groups", ["a", "b", "c", ["a", "c"], ["a", "b", "c"]]
-)
-def test_group_by_pandas_compat(groups):
-    with cudf.option_context("mode.pandas_compatible", True):
-        df = cudf.DataFrame(
-            {
-                "a": [1, 3, 2, 3, 3],
-                "b": ["x", "a", "y", "z", "a"],
-                "c": [10, 13, 11, 12, 12],
-            }
-        )
-        pdf = df.to_pandas()
-
-        assert_eq(pdf.groupby(groups).max(), df.groupby(groups).max())
-
-
-class TestSample:
-    @pytest.fixture(params=["default", "rangeindex", "intindex", "strindex"])
-    def index(self, request):
-        n = 12
-        if request.param == "rangeindex":
-            return cudf.RangeIndex(2, n + 2)
-        elif request.param == "intindex":
-            return cudf.Index(
-                [2, 3, 4, 1, 0, 5, 6, 8, 7, 9, 10, 13], dtype="int32"
-            )
-        elif request.param == "strindex":
-            return cudf.Index(list(string.ascii_lowercase[:n]))
-        elif request.param == "default":
-            return None
-
-    @pytest.fixture(
-        params=[
-            ["a", "a", "b", "b", "c", "c", "c", "d", "d", "d", "d", "d"],
-            [1, 1, 2, 2, 3, 3, 3, 4, 4, 4, 4, 4],
-        ],
-        ids=["str-group", "int-group"],
-    )
-    def df(self, index, request):
-        return cudf.DataFrame(
-            {"a": request.param, "b": request.param, "v": request.param},
-            index=index,
-        )
-
-    @pytest.fixture(params=["a", ["a", "b"]], ids=["single-col", "two-col"])
-    def by(self, request):
-        return request.param
-
-    def expected(self, df, *, n=None, frac=None):
-        value_counts = collections.Counter(df.a.values_host)
-        if n is not None:
-            values = list(
-                itertools.chain.from_iterable(
-                    itertools.repeat(v, n) for v in value_counts.keys()
-                )
-            )
-        elif frac is not None:
-            values = list(
-                itertools.chain.from_iterable(
-                    itertools.repeat(v, round(count * frac))
-                    for v, count in value_counts.items()
-                )
-            )
-        else:
-            raise ValueError("Must provide either n or frac")
-        values = cudf.Series(sorted(values), dtype=df.a.dtype)
-        return cudf.DataFrame({"a": values, "b": values, "v": values})
-
-    @pytest.mark.parametrize("n", [None, 0, 1, 2])
-    def test_constant_n_no_replace(self, df, by, n):
-        result = df.groupby(by).sample(n=n).sort_values("a")
-        n = 1 if n is None else n
-        assert_eq(self.expected(df, n=n), result.reset_index(drop=True))
-
-    def test_constant_n_no_replace_too_large_raises(self, df):
-        with pytest.raises(ValueError):
-            df.groupby("a").sample(n=3)
-
-    @pytest.mark.parametrize("n", [1, 2, 3])
-    def test_constant_n_replace(self, df, by, n):
-        result = df.groupby(by).sample(n=n, replace=True).sort_values("a")
-        assert_eq(self.expected(df, n=n), result.reset_index(drop=True))
-
-    def test_invalid_arguments(self, df):
-        with pytest.raises(ValueError):
-            df.groupby("a").sample(n=1, frac=0.1)
-
-    def test_not_implemented_arguments(self, df):
-        with pytest.raises(NotImplementedError):
-            # These are valid weights, but we don't implement this yet.
-            df.groupby("a").sample(n=1, weights=[1 / len(df)] * len(df))
-
-    @pytest.mark.parametrize("frac", [0, 1 / 3, 1 / 2, 2 / 3, 1])
-    @pytest.mark.parametrize("replace", [False, True])
-    def test_fraction_rounding(self, df, by, frac, replace):
-        result = (
-            df.groupby(by).sample(frac=frac, replace=replace).sort_values("a")
-        )
-        assert_eq(self.expected(df, frac=frac), result.reset_index(drop=True))
-
-
-class TestHeadTail:
-    @pytest.fixture(params=[-3, -2, -1, 0, 1, 2, 3], ids=lambda n: f"{n=}")
-    def n(self, request):
-        return request.param
-
-    @pytest.fixture(
-        params=[False, True], ids=["no-preserve-order", "preserve-order"]
-    )
-    def preserve_order(self, request):
-        return request.param
-
-    @pytest.fixture
-    def df(self):
-        return cudf.DataFrame(
-            {
-                "a": [1, 0, 1, 2, 2, 1, 3, 2, 3, 3, 3],
-                "b": [0, 1, 2, 4, 3, 5, 6, 7, 9, 8, 10],
-            }
-        )
-
-    @pytest.fixture(params=[True, False], ids=["head", "tail"])
-    def take_head(self, request):
-        return request.param
-
-    @pytest.fixture
-    def expected(self, df, n, take_head, preserve_order):
-        if n == 0:
-            # We'll get an empty dataframe in this case
-            return df._empty_like(keep_index=True)
-        else:
-            if preserve_order:
-                # Should match pandas here
-                g = df.to_pandas().groupby("a")
-                if take_head:
-                    return g.head(n=n)
-                else:
-                    return g.tail(n=n)
-            else:
-                # We groupby "a" which is the first column. This
-                # possibly relies on an implementation detail that for
-                # integer group keys, cudf produces groups in sorted
-                # (ascending) order.
-                keyfunc = operator.itemgetter(0)
-                if take_head or n == 0:
-                    # Head does group[:n] as does tail for n == 0
-                    slicefunc = operator.itemgetter(slice(None, n))
-                else:
-                    # Tail does group[-n:] except when n == 0
-                    slicefunc = operator.itemgetter(
-                        slice(-n, None) if n else slice(0)
-                    )
-                values_to_sort = np.hstack(
-                    [df.values_host, np.arange(len(df)).reshape(-1, 1)]
-                )
-                expect_a, expect_b, index = zip(
-                    *itertools.chain.from_iterable(
-                        slicefunc(list(group))
-                        for _, group in itertools.groupby(
-                            sorted(values_to_sort.tolist(), key=keyfunc),
-                            key=keyfunc,
-                        )
-                    ),
-                    strict=True,
-                )
-                return cudf.DataFrame(
-                    {"a": expect_a, "b": expect_b}, index=index
-                )
-
-    def test_head_tail(self, df, n, take_head, expected, preserve_order):
-        if take_head:
-            actual = df.groupby("a").head(n=n, preserve_order=preserve_order)
-        else:
-            actual = df.groupby("a").tail(n=n, preserve_order=preserve_order)
-        assert_eq(actual, expected)
-
-
-def test_head_tail_empty():
-    # GH #13397
-
-    values = [1, 2, 3]
-    pdf = pd.DataFrame({}, index=values)
-    df = cudf.DataFrame({}, index=values)
-
-    expected = pdf.groupby(pd.Series(values)).head()
-    got = df.groupby(cudf.Series(values)).head()
-    assert_eq(expected, got, check_column_type=False)
-
-    expected = pdf.groupby(pd.Series(values)).tail()
-    got = df.groupby(cudf.Series(values)).tail()
-
-    assert_eq(expected, got, check_column_type=False)
-
-
-@pytest.mark.parametrize(
-    "groups", ["a", "b", "c", ["a", "c"], ["a", "b", "c"]]
-)
-@pytest.mark.parametrize("sort", [True, False])
-def test_group_by_pandas_sort_order(groups, sort):
-    with cudf.option_context("mode.pandas_compatible", True):
-        df = cudf.DataFrame(
-            {
-                "a": [10, 1, 10, 3, 2, 1, 3, 3],
-                "b": [5, 6, 7, 1, 2, 3, 4, 9],
-                "c": [20, 20, 10, 11, 13, 11, 12, 12],
-            }
-        )
-        pdf = df.to_pandas()
-
-        assert_eq(
-            pdf.groupby(groups, sort=sort).sum(),
-            df.groupby(groups, sort=sort).sum(),
-        )
-
-
-@pytest.mark.parametrize(
-    "dtype",
-    ["int32", "int64", "float64", "datetime64[ns]", "timedelta64[ns]", "bool"],
-)
-@pytest.mark.parametrize(
-    "reduce_op",
-    [
-        "min",
-        "max",
-        "idxmin",
-        "idxmax",
-        "first",
-        "last",
-    ],
-)
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="Fails in older versions of pandas",
-)
-def test_group_by_empty_reduction(dtype, reduce_op):
-    gdf = cudf.DataFrame({"a": [], "b": [], "c": []}, dtype=dtype)
-    pdf = gdf.to_pandas()
-
-    gg = gdf.groupby("a")["c"]
-    pg = pdf.groupby("a")["c"]
-
-    assert_eq(
-        getattr(gg, reduce_op)(), getattr(pg, reduce_op)(), check_dtype=True
-    )
 
 
 @pytest.mark.parametrize(
